Document ams-visually-hidden as utility class
