+++ storybook/src/utils/VisuallyHidden/VisuallyHidden.docs.mdx
@@ -1,0 +1,8 @@
+import { Controls, Markdown, Primary } from "@storybook/blocks";
+import README from "../../../../packages/css/src/components/visually-hidden/README.md?raw";
+
+<Markdown>{README}</Markdown>
+
+<Primary />
+
+<Controls />
